[fix] インターフェース、型の定義場所修正
