+++ typescript/apps/admin/components/advanced/snackbar-container/index.vue
@@ -1,7 +1,7 @@
 <template>
   <transition-group tag="div" name="snackbar" class="snackbar-container d-flex flex-column pa-5 align-end">
     <template v-for="s of snackbarList">
-      <v-card :key="s.id" :value="true" :color="s.type" width="400px" class="snackbar-item">
+      <v-card :key="s.id" :color="s.type" width="400px" class="snackbar-item">
         <v-card-text class="white--text font-weight-bold">
           {{ s.message }}
         </v-card-text>
@@ -49,11 +49,7 @@
 <script lang="ts">
 import { Vue, Component } from 'vue-property-decorator'
 import { v4 as uuidv4 } from 'uuid';
-import { SnackbarData } from './types'
-
-export interface ISnackbarConteiner {
-  addSnackbar(data: Omit<SnackbarData, 'id' | 'duration'>, duration?: number): void
-}
+import { ISnackbarConteiner, SnackbarData } from '~/types/snackbar-container'
 
 @Component({})
 export default class SnackbarContainer extends Vue implements ISnackbarConteiner {
