--- typescript/apps/admin/components/advanced/snackbar-container/types.ts
@@ -1,6 +1,0 @@
-export type SnackbarData = {
-  id: string
-  type: 'success' | 'error'
-  message: string
-  duration: number
-}
