+++ typescript/apps/admin/types/snackbar-container.ts
@@ -1,0 +1,10 @@
+export type SnackbarData = {
+  id: string
+  type: 'success' | 'error'
+  message: string
+  duration: number
+}
+
+export interface ISnackbarConteiner {
+  addSnackbar(data: Omit<SnackbarData, 'id' | 'duration'>, duration?: number): void
+}
